requirements.txt
Django==4.2.16
djangorestframework==3.14.0
python-decouple==3.8

# MongoDB
mongoengine==0.27.0
pymongo==4.6.0

# WebSockets (Django Channels)
channels==4.0.0
daphne==4.0.0
channels-redis==4.1.0

# CORS
django-cors-headers==4.3.1

# Authentication & Security
PyJWT==2.8.0
bcrypt==4.1.2
pyotp==2.9.0
qrcode[pil]==7.4.2
cryptography==42.0.5
bleach==6.1.0

# Google APIs
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.119.0
google-generativeai==0.3.1

# OpenAI GPT (Primary AI)
openai>=1.30.0

# Email & SMS
twilio==8.11.0

# PDF Generation
reportlab==4.0.7

# Resume Parsing
pdfminer.six==20221105
python-docx==1.1.0

# Performance Monitoring
psutil==5.9.6

# Utilities
Pillow==10.2.0
python-dateutil==2.8.2
